General#network2hostnamesharepath was broken
if a local drive from a usb or external hardrive is attached then the original
method tried to find a network path even if it didn't exist.  This kept on
making hostname_share undefined and thus would raise an exception.

I changed it to first get all network drives into an array.  Then seeing if the
drive letter is in the array would let me know if was local or network drive